requirements.txt
mysql-connector-python>=2.0.3
Pygments>=2.0.2,<3.0
networkx>=2.5
APScheduler>=3.0.2,<4.0
pandas>=0.16.0
psycopg2>=2.9.1
python-dateutil>=2.4.2,<3.0
requests>=2.26.0
GitPython>=3.1.18
